+++ src/app/hooks/useLazyRequest.js
@@ -1,4 +1,4 @@
-import { useState } from 'react';
+import { useState, useCallback } from 'react';
 
 import { errorApiHandler } from '../../utils/errorHandler';
 
@@ -7,18 +7,22 @@ const useLazyRequest = ({ request }) => {
   const [submitError, setSubmitError] = useState([]);
   const [response, setResponse] = useState();
 
-  const sendRequest = dataPayload => {
-    setIsLoading(true);
-    request(dataPayload).then(data => {
-      if (data && data.ok) {
-        setSubmitError(null);
-      } else {
-        setSubmitError(errorApiHandler(data));
-      }
-      setResponse(data);
-      setIsLoading(false);
-    });
-  };
+  const sendRequest = useCallback(
+    dataPayload => {
+      setIsLoading(true);
+      request(dataPayload).then(data => {
+        if (data && data.ok) {
+          setSubmitError(null);
+        } else {
+          setSubmitError(errorApiHandler(data));
+        }
+        setResponse(data);
+        setIsLoading(false);
+      });
+    },
+    [request]
+  );
+
   return [isLoading, submitError, response, sendRequest];
 };
