+++ src/components/Router/index.js
@@ -3,7 +3,7 @@ import { BrowserRouter, Switch } from 'react-router-dom';
 
 import SignupForm from '../../screens/SignupForm';
 import LoginForm from '../../screens/LoginForm';
-import Books from '../../screens/Books';
+import BookDetail from '../../screens/BookDetail';
 import Home from '../../screens/Home';
 import HandleRoute from '../HandleRoute';
 import { PATHS } from '../../constants/paths';
@@ -12,7 +12,7 @@ function Router() {
   return (
     <BrowserRouter>
       <Switch>
-        <HandleRoute path={PATHS.bookDetail} component={Books} isPrivate />
+        <HandleRoute path={PATHS.bookDetail} component={BookDetail} isPrivate />
         <HandleRoute path={PATHS.login} component={LoginForm} />
         <HandleRoute path={PATHS.signup} component={SignupForm} />
         <HandleRoute exact path={PATHS.root} component={Home} isPrivate />
