--- src/contexts/bookDetailContext.js
@@ -1,3 +1,0 @@
-import React from 'react';
-
-export const BookDetailContext = React.createContext();
