+++ src/screens/BookDetail/i18n.js
