+++ src/screens/BookDetail/styles.module.scss
@@ -17,7 +17,6 @@
   color: $black;
   font-size: 30px;
   font-weight: 700;
-  margin-bottom: 20px;
   padding-bottom: 20px;
 }
