--- src/screens/Books/i18n.js
@@ -1,5 +1,0 @@
-import i18next from 'i18next';
-
-i18next.addResources('es', 'Home', {
-  linkLogout: 'Logout'
-});
